pyproject.toml
[project]
name = "brain-mcp"
version = "6.0.0"
description = "AI Eyes, User Hands — Live screen sharing assistant with Gemini"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "Pillow>=10.0",
    "pystray>=0.19",
    "websockets>=12.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
]

[project.scripts]
brain = "main_humphi:main"
